+++ WebApp/restart_bot.sh
