+++ WebApp/static/manifest.json
