+++ WebApp/templates/register.html
@@ -2,7 +2,7 @@
 <html lang="en">
     <head>
         <meta name="viewport" content="width=device-width, Initial-scale=1">
-        <title>Task App</title>
+        <title>Ma Liknot</title>
         <link rel="stylesheet" type="text/css" href="/static/login.css" media="screen">
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
     </head>
